// types/styles.ts
import { TextStyle, ViewStyle } from "react-native";

export interface Styles {
  container: ViewStyle;
  hero: ViewStyle;
  heroIconContainer: ViewStyle;
  heroIcon: TextStyle;
  heroTitle: TextStyle;
  heroTitleMobile: TextStyle;
  heroSubtitle: TextStyle;
  heroSubtitleMobile: TextStyle;
  heroBadges: ViewStyle;
  badge: ViewStyle;
  badgeText: TextStyle;
  tabContainer: ViewStyle;
  tab: ViewStyle;
  tabActive: ViewStyle;
  tabText: TextStyle;
  tabTextActive: TextStyle;
  infoBanner: ViewStyle;
  infoBannerIcon: TextStyle;
  infoBannerContent: ViewStyle;
  infoBannerTitle: TextStyle;
  infoBannerText: TextStyle;
  section: ViewStyle;
  sectionTitle: TextStyle;
  sectionTitleMobile: TextStyle;
  step: ViewStyle;
  stepMobile: ViewStyle;
  stepLeft: ViewStyle;
  stepNumber: ViewStyle;
  stepNumberSuccess: ViewStyle;
  stepNumberText: TextStyle;
  stepLine: ViewStyle;
  stepRight: ViewStyle;
  stepTitle: TextStyle;
  stepDescription: TextStyle;
  stepNote: TextStyle;
  storeButtons: ViewStyle;
  storeButtonsMobile: ViewStyle;
  storeButton: ViewStyle;
  storeButtonIos: ViewStyle;
  storeButtonAndroid: ViewStyle;
  storeButtonIcon: TextStyle;
  storeButtonLabel: TextStyle;
  storeButtonName: TextStyle;
  codeBlock: ViewStyle;
  codeHeader: ViewStyle;
  codeHeaderText: TextStyle;
  codeContent: ViewStyle;
  codeText: TextStyle;
  codeComment: TextStyle;
  qrInstructions: ViewStyle;
  qrInstruction: ViewStyle;
  qrInstructionIcon: TextStyle;
  qrInstructionContent: ViewStyle;
  qrInstructionTitle: TextStyle;
  qrInstructionText: TextStyle;
  alternativeSection: ViewStyle;
  alternativeTitle: TextStyle;
  alternativeText: TextStyle;
  featuresGrid: ViewStyle;
  featuresGridMobile: ViewStyle;
  featureCard: ViewStyle;
  featureCardMobile: ViewStyle;
  featureIconContainer: ViewStyle;
  featureIconLarge: TextStyle;
  featureTitle: TextStyle;
  featureDescription: TextStyle;
  techSection: ViewStyle;
  techTitle: TextStyle;
  techGrid: ViewStyle;
  techItem: ViewStyle;
  techLabel: TextStyle;
  techValue: TextStyle;
  footer: ViewStyle;
  footerText: TextStyle;
  footerCopyright: TextStyle;
}